+++ NEWS.md
@@ -1,6 +1,8 @@
 # heatwaveR 0.5.3.9000 (2024-03-27)
 
 - Basic hourly functionality added to `ts2clm()`
+- Tests added for `detect_event()`
+- Minor tweak to `exceedance()` to work with hourly data
 
 # heatwaveR 0.5.2.9009 (2024-01-23)